Dashboards: type v2 AnnotationPanelFilter.ids as number
Customer dashboards use panel ids in the int64 range (e.g. 23763571993).
The previous uint32 type on the v2 AnnotationPanelFilter.ids field made
the v1 to v2 conversion 500 with "cannot unmarshal number ... of type
uint32" for any annotation query that filtered by such a panel id.

Use CUE `number` (Go `[]float64`) across v2alpha1, v2beta1 and v2 to
match the existing `LibraryPanelKindSpec.id: number` declaration in the
same spec. This avoids the OpenAPI `format: int64` string-polymorphism
quirk and keeps the conversion code simple - panel ids decode from JSON
as float64 directly, no int64 casts needed. Add a regression test that
converts a v1 dashboard whose annotation panel-filter references
23763571993 and asserts both ids survive intact.

+++ apps/dashboard/pkg/apis/dashboard/v2/dashboard_spec_gen.go
@@ -88,14 +88,14 @@ type DashboardAnnotationPanelFilter struct {
 	// Should the specified panels be included or excluded
 	Exclude *bool `json:"exclude,omitempty"`
 	// Panel IDs that should be included or excluded
-	Ids []uint32 `json:"ids"`
+	Ids []float64 `json:"ids"`
 }
 
 // NewDashboardAnnotationPanelFilter creates a new DashboardAnnotationPanelFilter object.
 func NewDashboardAnnotationPanelFilter() *DashboardAnnotationPanelFilter {
 	return &DashboardAnnotationPanelFilter{
 		Exclude: (func(input bool) *bool { return &input })(false),
-		Ids:     []uint32{},
+		Ids:     []float64{},
 	}
 }
